fix: update dotfile installer (robust + reinstall)
This version is interactive, more robust, and allows for reinstallation.

On robustness: There was a previous issue with the move command where if
the destination did not have the parent directories then it would not
work. For some reason this worked on a minimal system but wouldn't work
on a dotfiles reinstall attempt. Now we explicitly create those parent
directories in the .dotfiles-backup directory before we move the files
there. Additionally, we use a more robust Bash 3 compatible method for doing this
instead of relying purely on egrep and xargs.